feat: cleanup and remove context + hooks
