+++ test/SliceZone.test.tsx
@@ -6,7 +6,6 @@ import { Slice } from "@prismicio/client";
 import { renderJSON } from "./__testutils__/renderJSON";
 
 import { SliceZone, TODOSliceComponent, SliceComponentProps } from "../src";
-import { SliceZoneResolver } from "../src/SliceZone";
 
 type StringifySliceComponentProps = {
 	/**
@@ -220,84 +219,6 @@ it("TODO component renders null in production", (ctx) => {
 	expect(consoleWarnSpy).not.toHaveBeenCalledWith(
 		expect.stringMatching(/could not find a component/i),
 		slices[1],
-	);
-
-	consoleWarnSpy.mockRestore();
-	process.env.NODE_ENV = originalNodeEnv;
-});
-
-// TODO: Remove in v3 when the `resolver` prop is removed.
-it("renders components from a resolver function for backwards compatibility with next-slicezone", async (ctx) => {
-	const slices = [
-		ctx.mock.value.slice(),
-		ctx.mock.value.slice(),
-		ctx.mock.value.slice(),
-	];
-	slices[0].slice_type = "foo_bar";
-	slices[1].slice_type = "barFoo";
-	slices[2].slice_type = "baz-qux";
-
-	const resolver: SliceZoneResolver<(typeof slices)[number]> = ({
-		sliceName,
-	}) => {
-		switch (sliceName) {
-			case "FooBar": {
-				return (props) => <StringifySliceComponent id="foo_bar" {...props} />;
-			}
-
-			case "BarFoo": {
-				return (props) => <StringifySliceComponent id="barFoo" {...props} />;
-			}
-
-			case "BazQux": {
-				return (props) => <StringifySliceComponent id="baz-qux" {...props} />;
-			}
-		}
-	};
-
-	const actual = renderJSON(<SliceZone slices={slices} resolver={resolver} />);
-	const expected = renderJSON(
-		<>
-			<StringifySliceComponent
-				id="foo_bar"
-				slice={slices[0]}
-				index={0}
-				slices={slices}
-				context={{}}
-			/>
-			<StringifySliceComponent
-				id="barFoo"
-				slice={slices[1]}
-				index={1}
-				slices={slices}
-				context={{}}
-			/>
-			<StringifySliceComponent
-				id="baz-qux"
-				slice={slices[2]}
-				index={2}
-				slices={slices}
-				context={{}}
-			/>
-		</>,
-	);
-
-	expect(actual).toStrictEqual(expected);
-});
-
-// TODO: Remove in v3 when the `resolver` prop is removed.
-it("logs a deprecation warning when using a resolver only in development", () => {
-	const originalNodeEnv = process.env.NODE_ENV;
-	process.env.NODE_ENV = "development";
-
-	const consoleWarnSpy = vi
-		.spyOn(console, "warn")
-		.mockImplementation(() => void 0);
-
-	renderJSON(<SliceZone slices={[]} resolver={() => void 0} />);
-
-	expect(consoleWarnSpy).toHaveBeenCalledWith(
-		expect.stringMatching(/the `resolver` prop is deprecated/i),
 	);
 
 	consoleWarnSpy.mockRestore();
