+++ test/index.test.ts
@@ -1,8 +1,8 @@
 import { it, expect } from "vitest";
-import * as prismicR from "@prismicio/richtext";
+import { Element as ClientElement } from "@prismicio/client/richtext";
 
-import * as prismicH from "../src";
+import { Element } from "../src";
 
-it("Element is an alias for @prismicio/richtext's Element", () => {
-	expect(prismicH.Element).toBe(prismicR.Element);
+it("Element is an alias for @prismicio/client/richtext's Element", () => {
+	expect(Element).toBe(ClientElement);
 });
