--- test/useAllPrismicDocumentsByUIDs.test.tsx
@@ -1,171 +1,0 @@
-// @vitest-environment happy-dom
-
-import { it, expect, afterEach } from "vitest";
-import * as React from "react";
-import * as prismic from "@prismicio/client";
-import { renderHook, cleanup, waitFor } from "@testing-library/react";
-import * as assert from "node:assert";
-
-import { createTestClient } from "./__testutils__/createTestClient";
-import { mockPrismicRestAPIV2 } from "./__testutils__/mockPrismicRestAPIV2";
-
-import { PrismicProvider, useAllPrismicDocumentsByUIDs } from "../src";
-
-afterEach(() => {
-	cleanup();
-});
-
-const createWrapper = (client: prismic.Client): React.ComponentType => {
-	// eslint-disable-next-line react/display-name
-	return (props) => <PrismicProvider client={client} {...props} />;
-};
-
-it("returns documents with every matching tag", async (ctx) => {
-	const client = createTestClient();
-
-	const model = ctx.mock.model.customType({
-		fields: {
-			uid: ctx.mock.model.uid(),
-		},
-	});
-	const docs = Array(3)
-		.fill(undefined)
-		.map(() => ctx.mock.value.document({ model }));
-	const uids = docs.map((doc) => doc.uid as string);
-	const type = model.id;
-
-	mockPrismicRestAPIV2({
-		ctx,
-		queryResponse: ctx.mock.api.query({
-			documents: docs,
-		}),
-		queryRequiredParams: {
-			q: [
-				`[${prismic.predicate.at("document.type", type)}]`,
-				`[${prismic.predicate.in(`my.${type}.uid`, uids)}]`,
-			],
-			pageSize: "100",
-		},
-	});
-
-	const { result } = renderHook(
-		() => useAllPrismicDocumentsByUIDs(type, uids),
-		{
-			wrapper: createWrapper(client),
-		},
-	);
-
-	await waitFor(() => {
-		assert.equal(result.current[1].state, "loaded");
-	});
-
-	expect(result.current[0]).toStrictEqual(docs);
-});
-
-it("supports params", async (ctx) => {
-	const client = createTestClient();
-
-	const model = ctx.mock.model.customType({
-		fields: {
-			uid: ctx.mock.model.uid(),
-		},
-	});
-	const docs = Array(3)
-		.fill(undefined)
-		.map(() => ctx.mock.value.document({ model }));
-	const uids = docs.map((doc) => doc.uid as string);
-	const type = model.id;
-
-	const params = {
-		pageSize: 2,
-	};
-
-	mockPrismicRestAPIV2({
-		ctx,
-		queryResponse: ctx.mock.api.query({
-			documents: docs,
-		}),
-		queryRequiredParams: {
-			q: [
-				`[${prismic.predicate.at("document.type", type)}]`,
-				`[${prismic.predicate.in(`my.${type}.uid`, uids)}]`,
-			],
-			pageSize: params.pageSize.toString(),
-		},
-	});
-
-	const { result } = renderHook(
-		() => useAllPrismicDocumentsByUIDs(type, uids, params),
-		{
-			wrapper: createWrapper(client),
-		},
-	);
-
-	await waitFor(() => {
-		assert.equal(result.current[1].state, "loaded");
-	});
-
-	expect(result.current[0]).toStrictEqual(docs);
-});
-
-it("supports explicit client", async (ctx) => {
-	const client = createTestClient();
-
-	const model = ctx.mock.model.customType({
-		fields: {
-			uid: ctx.mock.model.uid(),
-		},
-	});
-	const docs = Array(3)
-		.fill(undefined)
-		.map(() => ctx.mock.value.document({ model }));
-	const uids = docs.map((doc) => doc.uid as string);
-	const type = model.id;
-
-	mockPrismicRestAPIV2({
-		ctx,
-		queryResponse: ctx.mock.api.query({
-			documents: docs,
-		}),
-		queryRequiredParams: {
-			q: [
-				`[${prismic.predicate.at("document.type", type)}]`,
-				`[${prismic.predicate.in(`my.${type}.uid`, uids)}]`,
-			],
-			pageSize: "100",
-		},
-	});
-
-	const { result } = renderHook(() =>
-		useAllPrismicDocumentsByUIDs(type, uids, { client }),
-	);
-
-	await waitFor(() => {
-		assert.equal(result.current[1].state, "loaded");
-	});
-
-	expect(result.current[0]).toStrictEqual(docs);
-});
-
-it("returns failed state on error", async (ctx) => {
-	const client = createTestClient();
-
-	mockPrismicRestAPIV2({
-		ctx,
-		accessToken: "invalid-token",
-	});
-
-	const { result } = renderHook(
-		() => useAllPrismicDocumentsByUIDs("type", ["uid"]),
-		{
-			wrapper: createWrapper(client),
-		},
-	);
-
-	await waitFor(() => {
-		assert.equal(result.current[1].state, "failed");
-	});
-
-	expect(result.current[1].error).instanceOf(prismic.ForbiddenError);
-	expect(result.current[0]).toBe(undefined);
-});
